Setup only copy env file if not exist yet

+++ setup.sh
@@ -1,1 +1,3 @@
-cp .env.schema .env
+if [ ! -f ./.env ]; then
+  cp .env.schema .env
+fi
